adjust to partest 1.1.0's new mixed Java/Scala compilation
upgrades partest from 1.0.17 to 1.1.0

https://github.com/scala/scala-partest/pull/69
changed the mode for mixed compilation, which used to be

  1. scalac *.java *.scala -d o
  2. javac  *.java         -d o -cp o
  3. scalac        *.scala -d o -cp o

Now the third step is skipped. This required some adjustments to existing
tests.

  - t7014 is split in two groups, the fix is for separate compilation.
  - t7582 is also split. It tests inliner warnings when inling code that accesses
    Java-defined package-private code. Inlining from Java only works in separate
    compilation (no bytecode available in mixed compilation).
  - Java compiler warnings of "run" tests were not reported in the old scheme,
    now they are. Deprecation / unchecked warnings were removed from t6240, t8786,
    varargs.
  - t4788 required a .check file update to pass, which hints at a bug. I will
    re-open SI-4788 and investigate later.
