--- test/files/neg/t7014/t7014.scala
@@ -1,3 +1,0 @@
-package t7014
-
-import ThreadSafetyLevel.COMPLETELY_THREADSAFE // refer to annotation so it gets parsed
