+++ test/files/neg/t7014/t7014_2.scala
@@ -1,0 +1,3 @@
+package t7014
+
+import ThreadSafetyLevel_1.COMPLETELY_THREADSAFE // refer to annotation so it gets parsed
