+++ test/files/run/t4788.check
@@ -1,5 +1,5 @@
 Some(@Ljava/lang/Deprecated;())
 None
-None
-Some(@LCAnnotation;() // invisible)
+Some(@LSAnnotation;())
+Some(@LCAnnotation;())
 Some(@LRAnnotation;())
