+++ test/files/run/t7582/PackageProtectedJava_1.java
@@ -1,6 +1,6 @@
 package p1;
 
 // public class, protected method
-public class PackageProtectedJava {
+public class PackageProtectedJava_1 {
 	static final int protectedMethod() { return 1; }
 }
