+++ test/files/run/t7582b/InlineHolder_2.scala
@@ -1,6 +1,6 @@
 package p1 {
   object InlineHolder {
-    @inline def inlinable = p1.PackageProtectedJava.protectedMethod() + 1
+    @inline def inlinable = p1.PackageProtectedJava_1.protectedMethod() + 1
   }
 }
