+++ test/files/run/t7582b/PackageProtectedJava_1.java
@@ -1,6 +1,6 @@
 package p1;
 
 // protected class, public method
-class PackageProtectedJava {
+class PackageProtectedJava_1 {
 	public static final int protectedMethod() { return 1; }
 }
