+++ versions.properties
@@ -24,6 +24,6 @@ jline.version=2.14.3
 scala-asm.version=5.1.0-scala-1
 
 # external modules, used internally (not shipped)
-partest.version.number=1.0.17
+partest.version.number=1.1.0
 
 scalacheck.version.number=1.11.6
